Use half_thickness variable

+++ src/grid/grid.c
@@ -670,7 +670,7 @@ glPaintRectangle (CompScreen		  *s,
 
     /* set outline rect smaller to avoid damage issues */
     /* TODO: maybe get a better way of doing this */
-    int half_thickness = thickness * 0.5;
+    float half_thickness = thickness * 0.5;
     glVertex2f (rect.x1 + half_thickness, rect.y1 + half_thickness);
     glVertex2f (rect.x2 - half_thickness, rect.y1 + half_thickness);
     glVertex2f (rect.x2 - half_thickness, rect.y2 - half_thickness);
